aider: Added putting.webp image to the first screen for course and hole selection.
# Aider chat conversation:

USER: I want the images/putting.webp to be a part of the first screen there the users select the cource and the number of holes. can you include this in the code?

ASSISTANT: To include the `images/putting.webp` image on the first screen where users select the course and the number of holes, we need to modify the `pages/index.js` file.

Here are the steps:

1. Import the `Image` component from `next/image`.
2. Add the `putting.webp` image to the first screen.

Here are the *SEARCH/REPLACE* blocks:

pages/index.js
```python
<<<<<<< SEARCH
import React, { useState, useEffect } from 'react';
=======
import React, { useState, useEffect } from 'react';
import Image from 'next/image';
import puttingImage from '../public/images/putting.webp';
>>>>>>> REPLACE
```

pages/index.js
```python
<<<<<<< SEARCH
                    <div className="mb-4">
=======
                    <div className="mb-4 flex justify-center">
                        <Image src={puttingImage} alt="Putting" width={500} height={300} />
                    </div>
                    <div className="mb-4">
>>>>>>> REPLACE
```

This will ensure that the `putting.webp` image is displayed on the first screen where users select the course and the number of holes.